+++ Sources/Particles/BlockParticleManager.cpp
@@ -320,7 +320,8 @@ int BlockParticleManager::GetNumBlockParticles() const
 int BlockParticleManager::GetNumRenderableParticles(bool noWorldOffset)
 {
 	int numparticlesToRender = 0;
-	for (int i = 0; i < m_vpBlockParticlesList.size(); ++i)
+
+	for (size_t i = 0; i < m_vpBlockParticlesList.size(); ++i)
 	{
 		// If we are a emitter creation particle, don't render.
 		if (m_vpBlockParticlesList[i]->m_isCreateEmitters == true)
